ensure that the inline connection listener is executed first
In some versions of node.js (tested with v0.12.7) the http.Server class closes
the socket immediately when it sees that the request is not HTTP, before we
manage to send the xml! So we need to make sure that our connection listener is
executed _first_. To do so we remove them all and we re-add them.